Fixed forward sliding to not yet loaded slide
If sliding forward and target slide is not yet loaded
(it is loaded during this move) then canvas jumps to
it without animation. The cause was the code that fixes
position of slider on new slide insertion. There's no
need for this if we go forward, because insertion of
a slide after current doesn't shift it anywhere.