Try for an implementation of plugin support
very-very-very draft, just a try

Tried to implement slider number/bullets navigation as
a plugin. Added two events: init and go. The former is
triggered in the end of slider initialization. The
latter is triggered on go of the slider to a slide.

Looks good - it's working. Good plugin api & event
facilities are needed.